+++ src/OpenTelemetry.AutoInstrumentation/Instrumentations/Wcf/Client/WcfClientInitializer.cs
@@ -14,7 +14,7 @@
 // limitations under the License.
 // </copyright>
 
-#if NETFRAMEWORK
+using OpenTelemetry.Instrumentation.Wcf;
 
 namespace OpenTelemetry.AutoInstrumentation.Instrumentations.Wcf.Client;
 
@@ -33,10 +33,9 @@ internal static class WcfClientInitializer
     public static void Initialize(IChannelFactory channelFactory)
     {
         var behaviors = channelFactory.Endpoint.Behaviors;
-        if (!behaviors.Contains(typeof(WcfClientEndpointBehavior)))
+        if (!behaviors.Contains(typeof(TelemetryEndpointBehavior)))
         {
-            behaviors.Add(new WcfClientEndpointBehavior());
+            behaviors.Add(new TelemetryEndpointBehavior());
         }
     }
 }
-#endif