+++ src/OpenTelemetry.AutoInstrumentation/Instrumentations/Wcf/Client/WcfClientIntegration.cs
@@ -13,7 +13,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 // </copyright>
-#if NETFRAMEWORK
 
 using OpenTelemetry.AutoInstrumentation.CallTarget;
 
@@ -28,18 +27,8 @@ namespace OpenTelemetry.AutoInstrumentation.Instrumentations.Wcf.Client;
     methodName: WcfClientConstants.InitializeEndpointMethodName,
     returnTypeName: ClrNames.Void,
     parameterTypeNames: new[] { ClrNames.String, WcfClientConstants.EndpointAddressTypeName },
-    minimumVersion: WcfCommonConstants.MinVersion,
-    maximumVersion: WcfCommonConstants.MaxVersion,
-    integrationName: WcfClientConstants.IntegrationName,
-    type: InstrumentationType.Trace)]
-[InstrumentMethod(
-    assemblyName: WcfCommonConstants.ServiceModelAssemblyName,
-    typeName: WcfClientConstants.ChannelFactoryTypeName,
-    methodName: WcfClientConstants.InitializeEndpointMethodName,
-    returnTypeName: ClrNames.Void,
-    parameterTypeNames: new[] { ClrNames.String, WcfClientConstants.EndpointAddressTypeName, WcfClientConstants.ConfigurationTypeName },
-    minimumVersion: WcfCommonConstants.MinVersion,
-    maximumVersion: WcfCommonConstants.MaxVersion,
+    minimumVersion: WcfCommonConstants.Min4Version,
+    maximumVersion: WcfCommonConstants.Max4Version,
     integrationName: WcfClientConstants.IntegrationName,
     type: InstrumentationType.Trace)]
 [InstrumentMethod(
@@ -48,8 +37,8 @@ namespace OpenTelemetry.AutoInstrumentation.Instrumentations.Wcf.Client;
     methodName: WcfClientConstants.InitializeEndpointMethodName,
     returnTypeName: ClrNames.Void,
     parameterTypeNames: new[] { WcfClientConstants.ServiceEndpointTypeName },
-    minimumVersion: WcfCommonConstants.MinVersion,
-    maximumVersion: WcfCommonConstants.MaxVersion,
+    minimumVersion: WcfCommonConstants.Min4Version,
+    maximumVersion: WcfCommonConstants.Max4Version,
     integrationName: WcfClientConstants.IntegrationName,
     type: InstrumentationType.Trace)]
 [InstrumentMethod(
@@ -58,10 +47,54 @@ namespace OpenTelemetry.AutoInstrumentation.Instrumentations.Wcf.Client;
     methodName: WcfClientConstants.InitializeEndpointMethodName,
     returnTypeName: ClrNames.Void,
     parameterTypeNames: new[] { WcfClientConstants.BindingTypeName, WcfClientConstants.EndpointAddressTypeName },
-    minimumVersion: WcfCommonConstants.MinVersion,
-    maximumVersion: WcfCommonConstants.MaxVersion,
+    minimumVersion: WcfCommonConstants.Min4Version,
+    maximumVersion: WcfCommonConstants.Max4Version,
     integrationName: WcfClientConstants.IntegrationName,
     type: InstrumentationType.Trace)]
+#if NET6_0_OR_GREATER
+[InstrumentMethod(
+    assemblyName: WcfCommonConstants.ServiceModelPrimitivesAssemblyName,
+    typeName: WcfClientConstants.ChannelFactoryTypeName,
+    methodName: WcfClientConstants.InitializeEndpointMethodName,
+    returnTypeName: ClrNames.Void,
+    parameterTypeNames: new[] { ClrNames.String, WcfClientConstants.EndpointAddressTypeName },
+    minimumVersion: WcfCommonConstants.Min6Version,
+    maximumVersion: WcfCommonConstants.Max6Version,
+    integrationName: WcfClientConstants.IntegrationName,
+    type: InstrumentationType.Trace)]
+[InstrumentMethod(
+    assemblyName: WcfCommonConstants.ServiceModelPrimitivesAssemblyName,
+    typeName: WcfClientConstants.ChannelFactoryTypeName,
+    methodName: WcfClientConstants.InitializeEndpointMethodName,
+    returnTypeName: ClrNames.Void,
+    parameterTypeNames: new[] { WcfClientConstants.ServiceEndpointTypeName },
+    minimumVersion: WcfCommonConstants.Min6Version,
+    maximumVersion: WcfCommonConstants.Max6Version,
+    integrationName: WcfClientConstants.IntegrationName,
+    type: InstrumentationType.Trace)]
+[InstrumentMethod(
+    assemblyName: WcfCommonConstants.ServiceModelPrimitivesAssemblyName,
+    typeName: WcfClientConstants.ChannelFactoryTypeName,
+    methodName: WcfClientConstants.InitializeEndpointMethodName,
+    returnTypeName: ClrNames.Void,
+    parameterTypeNames: new[] { WcfClientConstants.BindingTypeName, WcfClientConstants.EndpointAddressTypeName },
+    minimumVersion: WcfCommonConstants.Min6Version,
+    maximumVersion: WcfCommonConstants.Max6Version,
+    integrationName: WcfClientConstants.IntegrationName,
+    type: InstrumentationType.Trace)]
+#endif
+#if NETFRAMEWORK
+[InstrumentMethod(
+    assemblyName: WcfCommonConstants.ServiceModelAssemblyName,
+    typeName: WcfClientConstants.ChannelFactoryTypeName,
+    methodName: WcfClientConstants.InitializeEndpointMethodName,
+    returnTypeName: ClrNames.Void,
+    parameterTypeNames: new[] { ClrNames.String, WcfClientConstants.EndpointAddressTypeName, WcfClientConstants.ConfigurationTypeName },
+    minimumVersion: WcfCommonConstants.Min4Version,
+    maximumVersion: WcfCommonConstants.Max4Version,
+    integrationName: WcfClientConstants.IntegrationName,
+    type: InstrumentationType.Trace)]
+#endif
 public static class WcfClientIntegration
 {
     /// <summary>
@@ -80,4 +113,3 @@ public static class WcfClientIntegration
         return CallTargetReturn.GetDefault();
     }
 }
-#endif