+++ test/test-applications/integrations/TestApplication.Wcf.Client.DotNet/StatusResponse.cs
@@ -1,4 +1,4 @@
-// <copyright file="IKeyedByTypeCollection.cs" company="OpenTelemetry Authors">
+// <copyright file="StatusResponse.cs" company="OpenTelemetry Authors">
 // Copyright The OpenTelemetry Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -13,11 +13,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 // </copyright>
-namespace OpenTelemetry.AutoInstrumentation.Instrumentations.Wcf;
 
-internal interface IKeyedByTypeCollection
+using System.Runtime.Serialization;
+
+namespace TestApplication.Wcf.Client.DotNet;
+
+[DataContract(Namespace = "http://opentelemetry.io/")]
+public class StatusResponse
 {
-    void Add(object o);
-
-    bool Contains(Type t);
+    [DataMember]
+    public DateTimeOffset ServerTime { get; set; }
 }